+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/model/Item.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
